rename analyser to analyzer for consistency
