+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/components/analyzer_no_data_message.tsx
